Energy_Management/elprice/web_datastore/requirements.txt
fastapi
pydantic
uvicorn
aiofiles
paho-mqtt
